// src/lib/account-wasm/account_wasm.d.ts
export interface JsCall {
  contractAddress: Felt
  entrypoint: string
  calldata: Felt[]
}

export interface Policy {
  target: string
  method: string
}

export type Felts = JsFelt[]

export type JsFelt = Felt

export interface SessionMetadata {
  session: Session
  maxFee: Felt | null
  credentials: Credentials | null
  isRegistered: boolean
}

export interface Session {
  policies: Policy[]
  expiresAt: number
}
// ...
export class CartridgeSessionAccount {
  free(): void
  /**
   * @param {string} rpc_url
   * @param {JsFelt} signer
   * @param {JsFelt} address
   * @param {JsFelt} chain_id
   * @param {(JsFelt)[]} session_authorization
   * @param {Session} session
   * @returns {CartridgeSessionAccount}
   */
  static new(
    rpc_url: string,
    signer: JsFelt,
    address: JsFelt,
    chain_id: JsFelt,
    session_authorization: JsFelt[],
    session: Session
  ): CartridgeSessionAccount
  /**
   * @param {string} rpc_url
   * @param {JsFelt} signer
   * @param {JsFelt} address
   * @param {JsFelt} owner_guid
   * @param {JsFelt} chain_id
   * @param {Session} session
   * @returns {CartridgeSessionAccount}
   */
  static new_as_registered(
    rpc_url: string,
    signer: JsFelt,
    address: JsFelt,
    owner_guid: JsFelt,
    chain_id: JsFelt,
    session: Session
  ): CartridgeSessionAccount
  /**
   * @param {JsFelt} hash
   * @param {(JsCall)[]} calls
   * @returns {Promise<Felts>}
   */
  sign(hash: JsFelt, calls: JsCall[]): Promise<Felts>
  /**
   * @param {(JsCall)[]} calls
   * @returns {Promise<any>}
   */
  execute(calls: JsCall[]): Promise<any>
  /**
   * @param {(JsCall)[]} calls
   * @returns {Promise<any>}
   */
  execute_from_outside(calls: JsCall[]): Promise<any>
}